Refacto

+++ Assets/Scripts/Menus/ButtonsScript.cs
@@ -5,7 +5,6 @@ using UnityEngine.SceneManagement;
 
 public class ButtonsScript : MonoBehaviour
 {
-    [SerializeField] private GameObject _Settings;
     private string root;
 
     private void Start()
@@ -32,7 +31,8 @@ public class ButtonsScript : MonoBehaviour
         {
             foreach (var file in files)
             {
-                File.Delete(file);
+                if(!file.EndsWith(".log"))
+                    File.Delete(file);
             }
         }
         SceneManager.LoadSceneAsync(1);
